Move shared date filter components to metabase/querying
SearchFilterDateDisplay and SearchFilterDatePicker are generic
querying-date UI (built on metabase/querying + metabase/ui), reused by
the transforms run-filter bar. Move them into
metabase/querying/filters/components so non-search modules can keep
using them once search becomes a feature module.
